requirements.txt
fastapi
uvicorn
python-multipart
tinydb
python-dotenv
google-cloud-vision
google-generativeai
azure-storage-blob
pymupdf